Treat position parameters as ROS coordinate system
ROS and CARLA use the opposite direction of the Y-axis. Before this
change, ROS parameters were used directly to set CARLA coordinates of
the camera and then was transformed to ROS coordinates for /tf. Now
it's the opposite. Both parameters and /tf are in ROS coordinates.

Theoretically, we also should do the same for orientation parameters,
but I don't needed it right now, so I'm postponing this to the future.